+++ lib/gem_changelog_diff.rb
@@ -1,6 +1,9 @@
 # frozen_string_literal: true
 
 require_relative "gem_changelog_diff/version"
+require_relative "gem_changelog_diff/outdated_gem"
+require_relative "gem_changelog_diff/detector"
+require_relative "gem_changelog_diff/cli"
 
 module GemChangelogDiff
   class Error < StandardError; end